- Fixed all warning for entire solution - Defined glew static in imgui for opengl to remove all glew warnings
git-tfs-id: [https://dev.azure.com/timjustin/]$/Angazi;C292

+++ Framework/Graphics/Src/DebugUI.cpp
@@ -113,7 +113,7 @@ namespace
 		style.WindowBorderSize = 1;
 		style.ChildBorderSize = 1;
 		style.PopupBorderSize = 1;
-		style.FrameBorderSize = is3D;
+		style.FrameBorderSize = static_cast<float>(is3D);
 
 		style.WindowRounding = 3;
 		style.ChildRounding = 3;
@@ -122,7 +122,7 @@ namespace
 		style.GrabRounding = 3;
 
 #ifdef IMGUI_HAS_DOCK 
-		style.TabBorderSize = is3D;
+		style.TabBorderSize = static_cast<float>(is3D);
 		style.TabRounding = 3;
 
 		colors[ImGuiCol_DockingEmptyBg] = ImVec4(0.38f, 0.38f, 0.38f, 1.00f);
