+++ VGP242/08_HelloSolarSystem/GameState.cpp
@@ -85,7 +85,7 @@ void GameState::Initialize()
 
 void GameState::Terminate()
 {
-	for (int i = 0; i < mPlanetTextures.size(); i++)
+	for (size_t i = 0; i < mPlanetTextures.size(); i++)
 	{
 		mPlanetTextures[i].Terminate();
 		mMoonTextures[i].Terminate();
@@ -128,7 +128,7 @@ void GameState::Render()
 	mPixelShader.Bind();
 	mSampler.BindVS();
 
-	for (int i = 0; i <mPlanetTextures.size(); i++)
+	for (size_t i = 0; i <mPlanetTextures.size(); i++)
 	{
 		//Planets
 		auto matWorld = Matrix4::RotationY(mRotation*mRotationSpeed[i]);
